tests: move new step tests into a dedicated file for msc steps
